backend/requirements.txt
fastapi
uvicorn
python-dotenv
cohere
qdrant-client
pydantic
httpx
